requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
Pillow>=10.0.0
setuptools>=65.5.1
wheel>=0.38.4
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
psycopg2-binary
python-dotenv